Tools with multiple capabilities- #19
* Preliminary implementation of multitool use

* Remove redundant tool_name in the spec

* Prototype internet tool

* Remove previous ways of accessing Gemini models

* Improve mention logic and interstitial on generative_event.py

* tools experimentation

- merge artifacts and canvas
- possibly merge ai tools

* Add AI media tools and remove internet.py keeping it separate

* Update tool schema to have lists

* comment stuff

* Tool updates

* Remove standalone image generator entry

* Remove internet entry

* Concise error handling
Add reaction to acknowledge the bot is performing inference tasks

* cute indicator

* Add almost all multimodal models

* Update tools

* Faster github search

* Fix GitHub search trying to access nonexistent repository key in ISSUES/PR search type

* Update system instructions

+++ tools/bing_search.py
@@ -6,34 +6,50 @@ import discord
 # Function implementations
 class Tool:
     tool_human_name = "Browse with Bing"
-    tool_name = "bing_search"
     def __init__(self, method_send, discord_ctx, discord_bot):
         self.method_send = method_send
         self.discord_ctx = discord_ctx
         self.discord_bot = discord_bot
 
-        self.tool_schema = {
-            "name": self.tool_name,
-            "description": "Search and fetch latest information and pull videos with Bing.",
-            "parameters": {
-                "type": "OBJECT",
-                "properties": {
-                    "query": {
-                        "type": "STRING"
+        self.tool_schema = [ 
+            {
+                "name": "bing_search",
+                "description": "Search and fetch latest information and pull videos with Bing.",
+                "parameters": {
+                    "type": "OBJECT",
+                    "properties": {
+                        "query": {
+                            "type": "STRING"
+                        },
+                        "n_results": {
+                            "type": "INTEGER",
+                        },
+                        "show_youtube_videos": {
+                            "type": "BOOLEAN"
+                        }
                     },
-                    "n_results": {
-                        "type": "INTEGER",
+                    "required": ["query"]
+                }
+            },
+            {
+                "name": "url_extractor",
+                "description": "Extract URLs to summarize",
+                "parameters": {
+                    "type": "OBJECT",
+                    "properties": {
+                        "urls": {
+                            "type": "ARRAY",
+                            "items": {
+                                "type": "STRING"
+                            }
+                        }
                     },
-                    "show_youtube_videos": {
-                        "type": "BOOLEAN"
-                    }
-                },
-                "required": ["query"]
+                    "required": ["urls"]
+                }
             }
-        }
+        ]
 
-    
-    async def _tool_function(self, query: str, n_results: int = 10, show_youtube_videos: bool = False):
+    async def _tool_function_bing_search(self, query: str, n_results: int = 10, show_youtube_videos: bool = False):
         # Must not be 50
         if n_results > 50:
             n_results = 10
@@ -53,8 +69,6 @@ class Tool:
 
         # Make a request
         async with _session.get(_endpoint, headers=_header, params=_params) as _response:
-            _imsg = await self.method_send(f"🔍 Searching for **{query}**")
-
             # Raise an exception
             try:
                 _response.raise_for_status()
@@ -76,8 +90,6 @@ class Tool:
             "results": []
         }]
         for _results in _data:
-            await _imsg.edit(f"🔍 Reading **{_results["name"]}**")
-
             # Append the data
             _output[0]["results"].append({
                 "title": _results["name"],
@@ -91,23 +103,17 @@ class Tool:
             _endpoint_video = "https://api.bing.microsoft.com/v7.0/videos/search"
 
             async with _session.get(_endpoint_video, headers=_header, params=_params) as _response:
-                await _imsg.edit(f"🔍 Searching for **{query}** videos")
-
                 try:
                     _response.raise_for_status()
                 except aiohttp.ClientConnectionError:
                     _output[0]["video_results"] = "No video results found"
-                    await _imsg.delete()
                     return _output
                 
                 _data = (await _response.json())["value"]
                 if not _data:
                     _output[0]["video_results"] = "No video results found"
-                    await _imsg.delete()
                     return _output
-                
-                await _imsg.edit("📽️ Let me look some relevant videos for you")
-                
+                        
                 # Add additional guidelines
                 _output[0]["video_result_guidelines"] = "Same as the former guidelines but only rank relevant YouTube videos to the user!"
                 _output[0]["video_result_rules"] = "You can only choose one YouTube video and put it at the end of your responses so it will be displayed to the user better."
@@ -137,6 +143,49 @@ class Tool:
         # Add footer about Microsoft Privacy Statement
         _sembed.set_footer(text="Used Bing search tool to fetch results, https://www.microsoft.com/en-us/privacy/privacystatement")
         await self.method_send(embed=_sembed)
+        return _output
 
-        await _imsg.delete()
+
+    # URL Extractor
+    async def _tool_function_url_extractor(self, urls: list):
+        # Must be 5 or below else error out
+        if len(urls) > 5:
+            raise ValueError("URLs must be 10 or below")
+        
+        # check for the aiohttp client session
+        if not hasattr(self.discord_bot, "_aiohttp_main_client_session"):
+            raise Exception("aiohttp client session for get requests not initialized and URL extraction cannot continue, please check the bot configuration")
+        
+        _session: aiohttp.ClientSession = self.discord_bot._aiohttp_main_client_session
+
+        # Download the URLs
+        _output = []
+        _imsg = await self.method_send("🔍 Extracting URLs")
+        for _url in urls:
+            _imsg = await _imsg.edit(f"🔍 Extracting URL: **{_url}**")
+            async with _session.get(_url) as _response:
+                # Check if the response is successful
+                if _response.status != 200:
+                    _output.append({
+                        "url": _url,
+                        "content": f"Failed to fetch URL with status code {_response.status}"
+                    })
+
+                # Check if its a binary file which must error out
+                if "application/octet-stream" in _response.headers["Content-Type"] and "Content-Disposition" in _response.headers:
+                    _output.append({
+                        "url": _url,
+                        "content": "Binary file, cannot extract text"
+                    })
+
+                _data = await _response.text()
+                _output.append({
+                    "url": _url,
+                    "content": _data
+                })
+
+        # Delete the message
+        if _imsg:
+            await _imsg.delete()
+
         return _output
